Add 2024 - Day 4 Part 1
It's giving the wrong answer, I checked with someone else's script
which gave the right answer.  I don't actually have the star yet on
this one.

It's tested to heck, I'm not sure where the problem is.

My original attempt didn't work because the diagonals I was taking on
the matrix weren't relative to the position of the X which some of the
examples revealed.

However the examples both pass with this solution as do all of my
other examples trying to suss out the edge case.